+++ app/src/main/resources/templates/recruit/view.html
@@ -141,6 +141,7 @@
       btn.addEventListener("click", editComment);
   });
 
+  // summernote 사용 설정
   $(document).ready(function() {
     var options = new Object();
     options.height = 300;
